Fix bug in ice_restoring, indexing of trcrn_rest
Fix bug in ice_restoring with respect to indexing of trcrn_rest. This will change answers when ice restoring is turned on.

The CICE restoring feature is not generally used, and we currently have no tests to cover that feature. This bug was discovered as part of an ongoing effort to improve regional grid capability with additional PRs to come in the future. A full test suite run on derecho with the intel compiler confirmed this, no answers changed.