Add an option to be able to update ice concentration (from data assimilation or retrievals/products) (#1060)
Add an option to be able to update ice concentration (from data assimilation or retrievals/products) at restart.

- This feature is being ported from CICE4, that is used in RTOFS:
    - We use the sea ice concentration from data assimilation (DA), performed using RTOFS-DA.
    - In this (CICE6) version usage is turned ON (default: OFF) via restart_mod="adjust_aice".
    - See similar/same implementation in the old (CICE4) version of ice_restart.F90, search for insert_ssmi.
    - @awallcraft ported the code he wrote (and is used in CICE4) to CICE6.
    - This PR was originally sent to the NOAA-EMC/CICE fork and following @NickSzapiro-NOAA's recommendation, being redirected here; that PR is now closed.
    - All the comments and conversations (@NickSzapiro-NOAA and @awallcraft) can be read via https://github.com/NOAA-EMC/CICE/pull/104, however, for the sake of self-completeness

A brief text/description follows:
One of the main goals of a data assimilation (DA)/prediction system that strives to provide close to observed sea ice concentrations and ice edge is to be able to reconcile the differences in modeled and observed sea ice concentrations derived from space borne instruments. Posey et al., 2025 implemented such a method with CICE4, it is hereby ported to CICE6. By default this feature is turned OFF. To turn it ON, set restart_mod='adjust_aice' in ice_in (input) namelist file and provide sic.nc to be able to nudge the modeled ice concentration to that from data assimilated (or observed) concentration.  By default, restart_mod='none'.  There is also a test mode, restart_mod='adjust_aice_test' which modifies aice read on restart to the nearest 5/100ths just to test the algorithm.


- With this proposed addition, an ability to insert data assimilated sea ice concentration is provided. Implementation is similar/same as in CICE4; see https://github.com/NOAA-EMC/RTOFS_GLO/blob/develop/sorc/rtofs_hycom.fd/src_2.2.99DHMTi-dist2B_relo_cice_v4.0e/source/ice_restart.F90 ; search for insert_ssmi. @awallcraft ported the code he wrote (and is used in CICE4) to CICE6.
- There is an acknowledgement that the coupler and/or MOM6 will need some modifications, but that is beyond the scope of this (CICE) repository. Such work will be taken up in future.

Co-authored-by: Nicholas Szapiro <149816583+NickSzapiro-NOAA@users.noreply.github.com>
Co-authored-by: Nick Szapiro <Nick.Szapiro@noaa.gov>
Co-authored-by: Till Rasmussen <tar@dmi.dk>
Co-authored-by: apcraig <anthony.p.craig@gmail.com>